Refactor.
Split the code in several chunks to simplify.
Removed old code.
