+++ lib/assets/javascripts/cartodb3/components/background-importer/background-import-item-view.js
@@ -1,54 +1,98 @@
 var CoreView = require('backbone/core-view');
+var Notifier = require('../../editor/components/notifier/notifier');
 var UploadConfig = require('../../config/upload-config');
 var ErrorDetailsView = require('./error-details-view');
 var WarningsDetailsView = require('./warnings-details-view');
 var TwitterImportDetailsDialog = require('./twitter-import-details-view');
-var template = require('./background-import-item.tpl');
 
 /**
  *  Import item within background importer
  *
  */
 
+var DELEGATIONS = {
+  loading: require('./delegated-import-views/loading'),
+  failed: require('./delegated-import-views/failed'),
+  warning: require('./delegated-import-views/warnings'),
+  success: require('./delegated-import-views/success')
+};
+
 module.exports = CoreView.extend({
-
-  className: 'ImportItem',
-  tagName: 'li',
-
-  events: {
-    'click .js-abort': '_removeItem',
-    'click .js-show_error': '_showImportError',
-    'click .js-show_warnings': '_showImportWarnings',
-    'click .js-show_stats': '_showImportStats',
-    'click .js-close': '_removeItem'
-  },
-
   initialize: function (opts) {
     if (!opts.userModel) throw new Error('userModel is required');
     if (!opts.configModel) throw new Error('configModel is required');
     if (!opts.modals) throw new Error('modals is required');
+    if (!opts.importModel) throw new Error('importModel is required');
 
     this._userModel = opts.userModel;
     this._configModel = opts.configModel;
     this._modals = opts.modals;
+    this._importModel = opts.importModel;
+    this._showSuccessDetailsButton = opts.showSuccessDetailsButton;
+
+    this._notification = Notifier.addNotification({});
 
     this._initBinds();
   },
 
-  render: function () {
-    var upload = this.model.get('upload');
-    var imp = this.model.get('import');
+  _initBinds: function () {
+    this._importModel.on('change', this.updateNotification, this);
+    this._importModel.on('remove', this.clean, this);
+    this._notification.on('notification:close', this._closeHandler, this);
+    this._notification.on('notification:action', this._actionHandler, this);
+    this.add_related_model(this._importModel);
+    this.add_related_model(this._notification);
+  },
+
+  _closeHandler: function () {
+    this.trigger('remove', this.model, this);
+    this._importModel.pause();
+    Notifier.removeNotification(this._notification);
+    this.clean();
+  },
+
+  _actionHandler: function (action) {
+    if (action === 'show_errors') {
+      this._showImportError();
+    } else if (action === 'show_stats') {
+      this._showImportStats();
+    } else if (action === 'show_table') {
+      console.log('show me the table');
+    } else if (action === 'show_warnings') {
+      this._showImportWarnings();
+    }
+  },
+
+  _getStatus: function () {
+    var status = 'loading';
+    var failed = this._importModel.hasFailed();
+    var completed = this._importModel.hasCompleted();
+    var warnings = this._importModel.getWarnings();
+
+    if (failed) {
+      status = 'error';
+    } else if (completed && !warnings) {
+      status = 'success';
+    } else if (completed && warnings) {
+      status = 'warning';
+    }
+
+    return status;
+  },
+
+  _getInfo: function () {
+    var upload = this._importModel.get('upload');
+    var imp = this._importModel.get('import');
 
     var d = {
       name: '',
-      state: this.model.get('state'),
+      state: this._importModel.get('state'),
       progress: '',
       service: '',
-      step: this.model.get('step'),
-      url: '',
-      failed: this.model.hasFailed(),
-      completed: this.model.hasCompleted(),
-      warnings: this.model.getWarnings(),
+      step: this._importModel.get('step'),
+      failed: this._importModel.hasFailed(),
+      completed: this._importModel.hasCompleted(),
+      warnings: this._importModel.getWarnings(),
       showSuccessDetailsButton: this._showSuccessDetailsButton,
       tables_created_count: imp.tables_created_count
     };
@@ -85,26 +129,21 @@ module.exports = CoreView.extend({
     d.service = upload.service_name;
 
     // Progress
-    if (this.model.get('step') === 'upload') {
-      d.progress = this.model.get('upload').progress;
+    if (this._importModel.get('step') === 'upload') {
+      d.progress = this._importModel.get('upload').progress;
     } else {
       d.progress = (UploadConfig.uploadStates.indexOf(d.state) / UploadConfig.uploadStates.length) * 100;
     }
 
-    this.$el.html(template(d));
+    d.progress = d.progress.toFixed(2);
 
-    return this;
+    return d;
   },
 
-  _initBinds: function () {
-    this.model.bind('change', this.render, this);
-    this.model.bind('remove', this.clean, this);
-  },
-
-  _removeItem: function () {
-    this.trigger('remove', this.model, this);
-    this.model.pause();
-    this.clean();
+  updateNotification: function () {
+    var status = this._getStatus();
+    var delegated = DELEGATIONS[status];
+    delegated.call(this, this._importModel, this._notification, status, this._getInfo(), this._showDetails);
   },
 
   _showImportStats: function () {
@@ -113,7 +152,7 @@ module.exports = CoreView.extend({
       return new TwitterImportDetailsDialog({
         modalModel: modalModel,
         userModel: self._userModel,
-        model: self.model
+        model: self._importModel
       });
     });
     modal.show();
@@ -121,7 +160,6 @@ module.exports = CoreView.extend({
 
   _showImportError: function () {
     var self = this;
-
     var modal = this._modals.create(function (modalModel) {
       return new ErrorDetailsView({
         configModel: self._configModel,
@@ -135,7 +173,6 @@ module.exports = CoreView.extend({
 
   _showImportWarnings: function () {
     var self = this;
-
     var modal = this._modals.create(function (modalModel) {
       return new WarningsDetailsView({
         modalModel: modalModel,
