--- lib/assets/javascripts/cartodb3/components/background-importer/background-polling-header-title.tpl
@@ -1,11 +1,0 @@
-<% if (totalPollings === 1) { %>
-  <% if (imports > 0) { %>
-  <%- _t('components.background-importer.connecting') %>
- <% } %>
-  <% if (geocodings > 0) { %>
-  <%- _t('components.background-importer.geocoding') %>
-  <% } %>
-<%- _t('components.background-importer.dataset') %>
-<% } else { %>
-<%- _t('components.background-importer.working') %>
-<% } %>
