--- lib/assets/javascripts/cartodb3/components/background-importer/background-polling-header.tpl
@@ -1,3 +1,0 @@
-<div class="BackgroundPolling-headerBadge LayoutIcon">
-  <i class="CDB-IconFont CDB-IconFont-cloud BackgroundPolling-headerBadgeIcon js-icon"></i>
-</div>
