--- lib/assets/javascripts/cartodb3/components/background-importer/background-polling.tpl
@@ -1,3 +1,0 @@
-<div class="BackgroundPolling-body">
-  <ul class="BackgroundPolling-list js-list"></ul>
-</div>
