+++ lib/assets/javascripts/cartodb3/components/background-importer/delegated-import-views/failed.tpl
@@ -1,0 +1,1 @@
+<%- _t('components.background-importer.background-importer-item.error-connecting', { name: name }) %> <% if (service) { %> <%- _t('components.background-importer.background-importer-item.from') %> <%- service %> <% } %>
