+++ lib/assets/javascripts/cartodb3/components/background-importer/delegated-import-views/loading.tpl
@@ -1,0 +1,2 @@
+<%- state %> <%- name %> <% if (service && service != "twitter_search") { %> <%- _t('components.background-importer.background-importer-item.from') %> <%- service %> <% } %>
+    <%- progress %>%
