+++ lib/assets/javascripts/cartodb3/components/background-importer/delegated-import-views/success.tpl
@@ -1,0 +1,1 @@
+<%- name %> <% if (service && service != "twitter_search") { %> <%- _t('components.background-importer.background-importer-item.from') %> <%- service %> <% } %> <%- _t('components.background-importer.background-importer-item.completed') %>!
