+++ lib/assets/javascripts/cartodb3/components/background-importer/delegated-import-views/warnings.tpl
@@ -1,0 +1,1 @@
+Some warnings were produced for <%- name %> <% if (service) { %> <%- _t('components.background-importer.background-importer-item.from') %> <%- service %> <% } %>
