+++ lib/assets/javascripts/cartodb3/components/background-importer/twitter-import-details-view.js
@@ -22,6 +22,7 @@ module.exports = Backbone.View.extend({
 
     var userTwitterValues = this._userModel.get('twitter');
     var availableTweets = userTwitterValues.quota - userTwitterValues.monthly_use;
+
     var url = vis && encodeURI(vis.viewUrl(this._userModel).edit()) || '';
 
     this.$el.html(
