--- lib/assets/javascripts/cartodb3/components/importer-manager/background-import-limit.tpl
@@ -1,5 +1,0 @@
-<% if (isUpgradeable) { %>
-<%- _t('components.background-importer.background-import-limit.hurry', { upgradeUrl: upgradeUrl }) %>
-<% } else { %>
-<%- _t('components.background-importer.background-import-limit.one-file', { importQuota: importQuota }) %>
-<% } %>
