+++ lib/assets/javascripts/cartodb3/editor.js
@@ -29,7 +29,7 @@ var ModalsServiceModel = require('./components/modals/modals-service-model');
 var viewFactory = require('./components/view-factory');
 var UserModel = require('./data/user-model');
 var EditorModel = require('./data/editor-model');
-var ImporterManager = require('./components/importer-manager/background-importer');
+var ImporterManager = require('./components/background-importer/background-importer');
 var BackgroundPollingModel = require('./data/editor-background-polling-model');
 var StyleManager = require('./editor/style/style-manager');
 var DeepInsightsIntegrations = require('./deep-insights-integrations');
