+++ lib/assets/test/spec/cartodb3/components/background-importer/background-import-item.spec.js
@@ -1,9 +1,9 @@
 var UserModel = require('../../../../../javascripts/cartodb3/data/user-model');
 var ConfigModel = require('../../../../../javascripts/cartodb3/data/config-model');
 var ImportsModel = require('../../../../../javascripts/cartodb3/data/background-importer/imports-model');
-var BackgroundImportItemView = require('../../../../../javascripts/cartodb3/components/importer-manager/background-import-item-view');
+var BackgroundImportItemView = require('../../../../../javascripts/cartodb3/components/background-importer/background-import-item-view');
 
-describe('editor/components/importer-manager/background-import-item', function () {
+describe('editor/components/background-importer/background-import-item', function () {
   beforeEach(function () {
     this.userModel = new UserModel({
       username: 'pepe',
@@ -24,9 +24,8 @@ describe('editor/components/importer-manager/background-import-item', function (
     });
 
     spyOn(this.model, 'bind').and.callThrough();
-
-    spyOn(BackgroundImportItemView.prototype, 'updateNotification');
-    spyOn(BackgroundImportItemView.prototype, 'clean');
+    spyOn(BackgroundImportItemView.prototype, 'updateNotification').and.callThrough();
+    spyOn(BackgroundImportItemView.prototype, 'clean').and.callThrough();
 
     this.importerView = new BackgroundImportItemView({
       modals: {},
@@ -35,6 +34,8 @@ describe('editor/components/importer-manager/background-import-item', function (
       userModel: this.userModel,
       configModel: this.configModel
     });
+
+    spyOn(this.importerView, '_getStatus').and.returnValue('success');
   });
 
   it('should create notification properly', function () {
@@ -55,10 +56,10 @@ describe('editor/components/importer-manager/background-import-item', function (
   it('should show display_name when it is available from import model', function () {
     this.model._importModel.set({ item_queue_id: 'hello-id', type: '' });
     this.model.set('state', 'importing');
-    expect(this.importerView._getInfo()).toContain('hello-id');
+    expect(this.importerView._notification.get('info')).toContain('hello-id');
     this.model._importModel.set('display_name', 'table_name_test');
     this.model.set('state', 'geocoding');
-    expect(this.importerView._getInfo()).toContain('table_name_test');
-    expect(this.importerView._getInfo()).not.toContain('hello-id');
+    expect(this.importerView._notification.get('info')).toContain('table_name_test');
+    expect(this.importerView._notification.get('info')).not.toContain('hello-id');
   });
 });
