+++ lib/assets/test/spec/cartodb3/components/background-importer/background-importer.spec.js
@@ -2,14 +2,14 @@ var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 var _ = require('underscore');
 var ImportsModel = require('../../../../../javascripts/cartodb3/data/background-importer/imports-model');
-var BackgroundImporter = require('../../../../../javascripts/cartodb3/components/importer-manager/background-importer');
+var BackgroundImporter = require('../../../../../javascripts/cartodb3/components/background-importer/background-importer');
 var BackgroundPollingModel = require('../../../../../javascripts/cartodb3/data/background-importer/background-polling-model');
 var ImportsCollection = require('../../../../../javascripts/cartodb3/data/background-importer/background-importer-imports-collection');
 var GeocodingsCollection = require('../../../../../javascripts/cartodb3/data/background-importer/background-importer-geocodings-collection');
 var UserModel = require('../../../../../javascripts/cartodb3/data/user-model');
 var ConfigModel = require('../../../../../javascripts/cartodb3/data/config-model');
 
-describe('editor/components/importer-manager/background-importer', function () {
+describe('editor/components/background-importer/background-importer', function () {
   beforeEach(function () {
     this.userModel = new UserModel({
       username: 'pepe',
